pkg/operator/reporters/blockers: Report serious keywords
Draw team attention to open bugs with keywords that move them into
important classes, mostly related to blocking other teams.
UpgradeBlocker isn't quite an other-team blocker, but it usually means
that we need to sort out an impact statement like [1] to determine if
we need to remove update recommendations in the production update
service [2].

I've placed the classifier for this ahead of the LifecycleStale guard
because we don't want stale bugs with these keywords.  A bug with both
LifecycleStale and a serious keyword should be addressed by removing
either LifecycleStale ("we should be working on this serious thing")
or the the serious keywords ("this is not actually as serious as the
keyword setter had thought").

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1861383#c7
[2]: https://github.com/openshift/cincinnati-graph-data/tree/f1e3897e3649e55c2564f6c2c9cf702179462e97#block-edges